Build the basic architecture for smcpaks. The concept is as follows: A Package can either be a local one (i.e. it refers to a .smcpak file) or a remote one that referes to a repository. Repositories can either be remote themselves if they point to an URL or local in case they point to a directory on the local hard disk (this is usually the case for the local SMC installation). Note that packages residing in a local repository are considered remote either, because this allows a similar treating of them as if they were in remote repositories, simplifying the process of setting up a remote repository from a local one (although this is not yet implemented).
The most parts of the code in this commit is untested, because I
didn't had a repository right now to test with (I think I'll build
up one on localhost), but creating, inspecting and decompressing
of packages already works. In order to create a smcpak, create
a directory obeying the smcpak package architecture (i.e. levels
in the "levels" subdirectory, music in "music", etc., and a
package specifiction named after the toplevel directory) and then
do:
  require_relative "lib/smc_get"
  SmcGet.setup
  pkg = SmcGet::Package.create("mypackage") #You've got a package now!
  puts pkg.location #=> /path/to/the/package
  puts pkg.spec.levels #=> All the levels...
  Dir.mkdir("abc")
  pkg.decompress("abc") #Decompress the package into abc